Rename a bunch of message types
This condenses a lot of the {Channel,Reactor}{Event,Reply,Message} soup.
All messages between components are of the Message enum, which (for now)
has Hangup and Text variants. The message enum is generic, so you can
send either events or replies. This makes all of the code significantly
less complicated and easier to read. (With only the very small price
that what used to be ChannelMessages now have a user, which will almost
always be None.)

The actual sequence of renames:

1. Make a new Message<T> enum
2. Rename ChannelMessage and ReactorMessage to Event
3. Rename ChannelEvent and ReactorEvent to Message<Event>
4. Rename ChannelReply and ReactorReply to Message<Reply>
5. Move ReactorSeed and ChannelSeed out of the hub
6. Delete the now-obsolete event module