Implement "does not compute" behavior
This really should have been two commits, but it was a pain to split it
up after and also nobody is reading this code. This adds a new Message
type, Ack(event_id, bool).

When the reactors get an event, they immediately ack it with either
yes, they will respond or no, they will not. When it receives an event,
the hub will note this event as pending: once all the reactors have
reported back, the hub will respond with a "Does not compute". This
required the hub to have a sender to its own reply channel, which was
nbd.