Rename a bunch of variables (part 2)
_rx and _tx were confusing: really, each component has an input and
output channel for talking to the hub. Reactors only talk to the hub, so
theirs are now called "input" and "output". Channels have an additional input
and output for talking to the world, so I called theirs from_hub and
to_hub. It would make sense for both of these names to be the same, but
I don't like the names from_hub and to_hub, so I didn't want to doubly
dislike them by sticking them in the reactors too.

Is this totally irrational? Yes. Do I care? Not so much.